Implement closeProjectBoardBtnClick and associated sub-functions
I created the closeProjectBoardBtnClick function to handle the submit click. Inside this function, I added two sub-functions: addProjectToProjectList (defined in projects.js) to process the logic for adding and removing projects, and removeProjectBoard to remove the project board layer. The render() function is called to re-render the page after any updates.